requirements.txt
APScheduler==3.9.1.post1
colorama==0.4.6
greenlet==2.0.1
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.45
tzdata==2022.7
tzlocal==4.2
